hourly user show service add
